requirements.txt
alembic==1.10.2
aniso8601==9.0.1
attrs==22.2.0
bidict==0.22.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
cryptography==39.0.2
docopt==0.6.2
exceptiongroup==1.1.1
Flask==2.2.2
Flask-SocketIO==5.3.2
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
h11==0.14.0
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
packaging==23.0
pipreqs==0.4.11
pluggy==1.0.0
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.6.0
pytest==7.2.2
flask_cors==3.0.10
python-decouple==3.8
python-dotenv==1.0.0
python-socketio==5.7.2
pytz==2022.7.1
psycopg2==2.9.5
python-dotenv==1.0.0
python-engineio==4.3.4
requests==2.28.2
simple-websocket==0.9.0
six==1.16.0
SQLAlchemy==2.0.4
tomli==2.0.1
typing_extensions==4.5.0
urllib3==1.26.14
Werkzeug==2.2.2
wsproto==1.2.0
yarg==0.1.9
